[esxcli-nvme-plugin] Integrate changeset 9211518
Remove placeholder URLs.

Starting 7.0, we are not publishing Bulletin specific KBs.
But, the URL fields on main and prod branch have retained
the example placeholder URLs that need to be cleaned up.
Please eliminate those and leave the KB URLs blank in the
bulletin metadata.

Originally submitted at: Wed, 8 Sep 2021 14:32:51 +0800

Signed-off-by: Wenchao Cui <wcui@vmware.com>