[esxcli-nvme-plugin] Integrate changeset 13358394
[esxcli-nvme-plugin] Add commands supported and effects log.
The commands supported and effects log page is used to describe
the commands that the controller supports and the effects of
those commands on the state of the NVM subsystem. As more and more
PCIe controllers support this log, add a command in esxcli-nvme-plugin
to retrieve this log page.

Originally submitted at: Wed May 22 18:06:00 2024 -0800