[esxcli-nvme-plugin] Integrate commit e0ac0663deaccac718d9c7e92ff0d8eb86099268
[esxcli-nvme-plugin] Modify Makefile to compile in DDK 9.0
To compile the esxcli-nvme-plugin in DDK 9.0, DEV_VIB_VERSION needs to
be defined in Makefile.
To run latest esxcli-nvme-plugin on old ESXi releases (< 9.1.0),
esxcliNvmePluginDom needs to be removed since it is not defined on these
ESXi releases.
This change also removes DVMK_DEVKIT_USES_BINARY_INCOMPATIBLE_APIS if
DDK version is not smaller than 8.0.3.

Originally submitted at: Tue Jun 3 10:00:18 2025 +0800