Remove unnecessary DMA constraint
This change removes the DMA constraint that requires buffer
sizes to be a multiple of 512. This has a side effect of causing
bounce buffers to be used when a buffer address is not a multiple
of 512.  Runtime testing shows that vSAN IOs are often only 8-byte
aligned, and the end result is a significant reduction of performance
compared to when this constraint isn't specified.